app/src/main/java/team27/healthe/GeoLocation.java
package team27.healthe;

public class GeoLocation {
    private double latitude;
    private double longitude;

    public GeoLocation(double lat, double lon) {
        latitude = lat;
        longitude = lon;
    }

    public double getLatitude() {
        return latitude;
    }

    public void setLatitude(double lat) {
        this.latitude = lat;
    }

    public double getLongitude() {
        return longitude;
    }

    public void setLongitude(double lon) {
        this.longitude = lon;
    }
}
